MS: Add CRT Generator.
The CRT fragment generator interfaces with the CRT hardware
indirectly through a backend DAQ program.  This program drops
files which the fragment generator reads as they are written.

This is a working module as far as reading data and returning
it goes, but a few changes are still needed to be functional
as part of ProtoDUNE:

* It must read the start-of-run time from the CRT rack's
timing board in order to normalized the timestamps of CRT
hits.

* It needs to automatically start and stop the backend DAQ, and
a convention is needed for where the files will be and how they
will be managed (i.e. deleted).

* [Probably more items that will be revealed upon testing.]