MS: Approach correct CRT timing interaction.
As per Alessandro, update (currently disabled) code that reads
registers from the timing board.  It now reads a register that is
available to us and does an atomic read of that one and the next,
which is what we will want to do with the run start time.  This
commit also adds notes about what is still wrong.