MS: Whitespace-and-similar cleanup in CRTInterface
* Remove all EOL whitespace.

* Reduce all line lengths to 88 characters or less, except for a few
that I'm about to make other changes to in another commit.  88
characters is the width of a terminal on my screen when I have two side
by side.  It's also close to the traditional limit of 80, so this is my
attempt to enforce a reasonable width without being too dogmatic.

* Remove one extraneous "return;" at the end of a void function.

* Be a little more consistent about when curly brackets are used for
one-line statements.  (I did not audit the whole file for this.)

* Removed some commented-out debugging log statements.