JCF: This is a de-facto (manual) reversion to 1a31e8cf60c3542d1eb1ab76a5de2af77839270f
I've modified dune-artdaq in this commit so that it's a carbon copy of commit 1a31e8cf60c3542d1eb1ab7
the reason for this is that the develop branch had been corrupted in the last couple of days via an i
which consisted of too many commits to revert one-by-one. 1a31e8cf60c3542d1eb1ab76a5de2af77839270f is
currently is the head of the local develop branch in /nfs/sw/work_dirs/dune-artdaq-June2019/srcs/dune
np04 cluster.