JCF: clone_and_rename_installation.sh can work with local build areas
It remains the case that clone_and_rename_installation.sh should only be
used with expert's permission, and only when needed - e.g., if there's a
problem with the network s.t. it's not possible to pull code from
Fermilab's central repository. However, should it ever be needed, it's
now the case that if the script determines that it's cloning a locally-built
dune-artdaq installation - i.e., one that resides on /scratch rather than
/nfs/sw/work_dirs - it will be able to handle this. I've successfully tested
dune-artdaq development areas created with this script (commenting out the exit
at the top) using run 8887 (a cloned local-build installation) and run 8888 (a
cloned, traditional-style NFS installation)